some modifications on startup code
added the option to choose which operation to be executed.
added text boxes for key and tap positions.
function encrypt image is implemented inside image operations class.

+++ ImageEncryptCompress/BitwiseOperations.cs
@@ -8,29 +8,29 @@ namespace ImageEncryptCompress
 {
     internal class BitwiseOperations
     {
-        public static int XORBits(int key, int numberOfBits, int tapPosition)
+        public static long XORBits(long key, int numberOfBits, int tapPosition)
         {
-            int keyMask = 1 << tapPosition;
-            int lastPosMask = 1 << numberOfBits - 1;
-            int capMask = (1 << numberOfBits) - 1;
-            int tapBit = (key & keyMask);
-            int lastBit = (key & lastPosMask) >> numberOfBits - tapPosition - 1;
-            int xorResult = (tapBit ^ lastBit) >> tapPosition;
-            int shiftedKey = key << 1;
-            int result = (shiftedKey | xorResult) & capMask;
+            long keyMask = 1 << tapPosition;
+            long lastPosMask = 1 << numberOfBits - 1;
+            long capMask = (1 << numberOfBits) - 1;
+            long tapBit = (key & keyMask);
+            long lastBit = (key & lastPosMask) >> numberOfBits - tapPosition - 1;
+            long xorResult = (tapBit ^ lastBit) >> tapPosition;
+            long shiftedKey = key << 1;
+            long result = (shiftedKey | xorResult) & capMask;
             return result;
         }
         public static int GeneratePassword(ref string key, int tapPosition, int numberOfBitsToGenerate = 8)
         {
-            int convertedKey = Convert.ToInt32(key, 2);
+            long convertedKey = Convert.ToInt32(key, 2);
             int numberOfBits = key.Length;
             for (int i = 0; i < numberOfBitsToGenerate; i++)
             {
                 convertedKey = XORBits(convertedKey, numberOfBits, tapPosition);
             }
             key = Convert.ToString(convertedKey, 2).PadLeft(numberOfBits, '0');
-            int result = convertedKey & ((1 << numberOfBitsToGenerate) - 1);
-            return result;
+            long result = convertedKey & ((1 << numberOfBitsToGenerate) - 1);
+            return (int)result;
         }
     }
 }
